% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{as.data.frame.compareClusterResult}
\alias{as.data.frame.compareClusterResult}
\title{Convert compareClusterResult to data frame}
\usage{
\method{as.data.frame}{compareClusterResult}(x, ...)
}
\arguments{
\item{x}{compareClusterResult object}

\item{...}{additional parameters}
}
\value{
data frame
}
\description{
Convert compareClusterResult to data frame
}
