% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{autofacet}
\alias{autofacet}
\title{Plotting utility functions for enrichplot package}
\usage{
autofacet(by = "row", scales = "free", levels = NULL)
}
\arguments{
\item{by}{one of 'row' or 'column'}

\item{scales}{whether 'fixed' or 'free'}

\item{levels}{set facet levels}
}
\value{
a ggplot object
}
\description{
This file contains plotting and visualization helper functions for enrichplot
Automatically split barplot or dotplot into several facets
}
