% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/barplot.R
\name{barplot}
\alias{barplot}
\alias{barplot,enrichResult-method}
\alias{barplot,gseaResult-method}
\alias{barplot,compareClusterResult-method}
\alias{barplot.enrichResult}
\alias{barplot.compareClusterResult}
\title{barplot}
\usage{
barplot(height, ...)

\S4method{barplot}{enrichResult}(height, ...)

\S4method{barplot}{gseaResult}(height, ...)

\S4method{barplot}{compareClusterResult}(height, ...)

\method{barplot}{enrichResult}(
  height,
  x = "Count",
  color = "p.adjust",
  showCategory = 8,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)

barplot.compareClusterResult(
  height,
  x = "Count",
  color = "p.adjust",
  showCategory = 5,
  by = "geneRatio",
  includeAll = TRUE,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)
}
\arguments{
\item{height}{enrichResult object}

\item{...}{other parameter, ignored}

\item{x}{one of 'Count' and 'GeneRatio'}

\item{color}{one of 'pvalue', 'p.adjust' and 'qvalue'}

\item{showCategory}{number of categories to show}

\item{font.size}{font size}

\item{title}{plot title}

\item{label_format}{a numeric value sets wrap length, alternatively a
custom function to format axis labels.
by default wraps names longer that 30 characters}

\item{by}{one of Count and GeneRatio}

\item{includeAll}{logical}
}
\value{
plot

ggplot object
}
\description{
barplot for enrichment result
}
\examples{
\dontrun{
    library(DOSE)
    data(geneList)
    de <- names(geneList)[1:100]
    x <- enrichDO(de)
    dotplot(x)
    # use `showCategory` to select the displayed terms. It can be a number of a vector of terms.
    barplot(x, showCategory = 10)
    categorys <- c("pre-malignant neoplasm", "intestinal disease",
                   "breast ductal carcinoma", "non-small cell lung carcinoma")
    barplot(x, showCategory = categorys)
}
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
barplot(x)
# use `showCategory` to select the displayed terms. It can be a number of a vector of terms.
barplot(x, showCategory = 10)
categorys <- c("pre-malignant neoplasm", "intestinal disease",
               "breast ductal carcinoma", "non-small cell lung carcinoma")
barplot(x, showCategory = categorys)
}
