% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{color_palette}
\alias{color_palette}
\title{Create color palette for continuous data}
\usage{
color_palette(colors)
}
\arguments{
\item{colors}{colors of length >=2}
}
\value{
color vector
}
\description{
Create color palette for continuous data
}
\examples{
color_palette(c("red", "yellow", "green"))
}
\author{
guangchuang yu
}
