% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot2}
\alias{dotplot2}
\title{dotplot2}
\usage{
dotplot2(object, x = "FoldEnrichment", vars = NULL, label = "auto", ...)
}
\arguments{
\item{object}{a compareClusterResult object}

\item{x}{selected variable to visualize in x-axis}

\item{vars}{selected Clusters to be compared, only length of two is supported}

\item{label}{to label the Clusters in the plot, should be a named vector}

\item{...}{additional parameters passed to dotplot}
}
\value{
a ggplot object
}
\description{
compare two clusters in the compareClusterResult object
}
\author{
Guangchuang Yu
}
