% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssplot.R
\name{dr}
\alias{dr}
\title{Dimension reduction}
\usage{
dr(enrichResult, drfun, drfunName, showCategory, split = NULL, pie = NULL)
}
\arguments{
\item{enrichResult}{Enrichment result}

\item{drfun}{Dimension reduction function}

\item{drfunName}{Name of dimension reduction function}

\item{showCategory}{A number or a vector of terms. If it is a number, 
the first n terms will be displayed. If it is a vector of terms, 
the selected terms will be displayed.}

\item{split}{separate result by 'category' variable}

\item{pie}{proportion of clusters in the pie chart, one of 'equal' (default) and 'Count'}
}
\description{
Dimension reduction
}
