% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/emapplot_cluster.r
\name{emapplot_cluster}
\alias{emapplot_cluster}
\alias{emapplot_cluster,enrichResult-method}
\alias{emapplot_cluster,gseaResult-method}
\alias{emapplot_cluster,compareClusterResult-method}
\alias{emapplot_cluster.enrichResult}
\alias{emapplot_cluster.compareClusterResult}
\title{emapplot_cluster}
\usage{
emapplot_cluster(x, showCategory = nrow(x), color = "p.adjust", ...)

\S4method{emapplot_cluster}{enrichResult}(x, showCategory = nrow(x), color = "p.adjust", ...)

\S4method{emapplot_cluster}{gseaResult}(x, showCategory = nrow(x), color = "p.adjust", ...)

\S4method{emapplot_cluster}{compareClusterResult}(x, showCategory = 30, color = "p.adjust", ...)

emapplot_cluster.enrichResult(
  x,
  showCategory = nrow(x),
  color = "p.adjust",
  line_scale = 0.1,
  with_edge = TRUE,
  method = "JC",
  nWords = 4,
  nCluster = NULL,
  split = NULL,
  min_edge = 0.2,
  cluster_label_scale = 1,
  semData = NULL,
  x_adjust = 0.5,
  y_adjust = 0.5,
  label_style = "shadowtext"
)

emapplot_cluster.compareClusterResult(
  x,
  showCategory = 30,
  color = "p.adjust",
  line_scale = 0.1,
  with_edge = TRUE,
  method = "JC",
  nWords = 4,
  nCluster = NULL,
  split = NULL,
  min_edge = 0.2,
  cluster_label_scale = 1,
  semData = NULL,
  x_adjust = 0.5,
  y_adjust = 0.2,
  pie = "equal",
  legend_n = 5,
  pie_scale = 1,
  label_style = "shadowtext"
)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{...}{additional parameters}

\item{line_scale}{scale of line width}

\item{with_edge}{if TRUE, draw the edges of the network diagram}

\item{method}{method of calculating the similarity between nodes, one of "Resnik",
"Lin", "Rel", "Jiang" , "Wang"  and "JC" (Jaccard similarity coefficient) methods}

\item{nWords}{the number of words in the cluster tags}

\item{nCluster}{the number of clusters}

\item{split}{separate result by 'category' variable}

\item{min_edge}{minimum percentage of overlap genes to display the edge, should between 0 and 1, default value is 0.2}

\item{cluster_label_scale}{scale of cluster labels size}

\item{semData}{GOSemSimDATA object}

\item{x_adjust, y_adjust}{If the tag distance of the two groups is less than (x_adjust, y_adjust), adjust them}

\item{label_style}{one of "shadowtext" and "ggforce"}

\item{pie}{proportion of clusters in the pie chart, one of 'equal' (default) or 'Count'}

\item{legend_n}{number of circle in legend}

\item{pie_scale}{scale of pie plot}
}
\value{
ggplot object
}
\description{
Functional grouping network diagram for enrichment result of
over-representation test or gene set enrichment analysis
}
\details{
This function visualizes gene sets as a grouped network (i.e. enrichment map).
Gene sets with high similarity tend to cluster together, making it easier for interpretation.
}
\examples{
\dontrun{
    library(clusterProfiler)
    library(org.Hs.eg.db)
    library(enrichplot)
    library(GOSemSim)
    library(DOSE)
    data(geneList)
       gene <- names(geneList)[abs(geneList) > 2]
       ego <- enrichGO(gene  = gene,
        universe      = names(geneList),
        OrgDb         = org.Hs.eg.db,
        ont           = "CC",
        pAdjustMethod = "BH",
        pvalueCutoff  = 0.01,
        qvalueCutoff  = 0.05,
        readable      = TRUE)
     d <- godata('org.Hs.eg.db', ont="BP")
     emapplot_cluster(ego, showCategory = 80, method = "Wang", semData = d)
   }
}
