% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{enrichplot-common-params}
\alias{enrichplot-common-params}
\title{Shared parameters for enrichment plots}
\arguments{
\item{color}{variable used to color enriched terms, e.g. \code{'pvalue'},
\code{'p.adjust'}, or \code{'qvalue'}.}

\item{showCategory}{number of categories to display, or a vector of terms.}

\item{size}{variable used to scale category size, one of \code{"geneRatio"},
\code{"Percentage"}, or \code{"count"}.}

\item{split}{apply \code{showCategory} to each category specified by \code{split},
e.g., \code{"ONTOLOGY"}, \code{"category"}, or \code{"intersect"}. Default is \code{NULL}.}

\item{font.size}{font size.}

\item{title}{figure title.}

\item{label_format}{a numeric wrap width, or a custom function to format
axis labels.}

\item{includeAll}{logical value.}
}
\description{
Shared parameters for enrichment plots
}
\keyword{internal}
