% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{enrichplot-term-params}
\alias{enrichplot-term-params}
\title{Shared term-plot parameters}
\arguments{
\item{showCategory}{number of categories to display, or a vector of terms.}

\item{color}{variable used to color enriched terms, e.g. \code{'pvalue'},
\code{'p.adjust'}, or \code{'qvalue'}.}

\item{label_format}{a numeric wrap width, or a custom function to format
axis labels.}
}
\description{
Shared term-plot parameters
}
\keyword{internal}
