% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\docType{data}
\name{enrichplot_point_shape}
\alias{enrichplot_point_shape}
\title{Predefined color palettes}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
enrichplot_point_shape
}
\description{
Predefined color palettes
}
\keyword{datasets}
