% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ep_str_wrap}
\alias{ep_str_wrap}
\title{ep_str_wrap internal string wrapping function}
\usage{
ep_str_wrap(string, width)
}
\arguments{
\item{string}{the string to be wrapped}

\item{width}{the maximum number of characters before wrapping to a new line}
}
\description{
ep_str_wrap internal string wrapping function
}
