% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-fortify.R
\name{fortify.enrichResult}
\alias{fortify.enrichResult}
\title{fortify}
\usage{
\method{fortify}{enrichResult}(model, data, showCategory = 5, by = "Count",
  order = FALSE, drop = FALSE, split = NULL, ...)
}
\arguments{
\item{model}{enrichResult object}

\item{data}{not use here}

\item{showCategory}{Category numbers to show}

\item{by}{one of Count and GeneRatio}

\item{order}{logical}

\item{drop}{logical}

\item{split}{separate result by 'split' variable}

\item{...}{additional parameter}
}
\description{
convert enrichResult object for ggplot2
}
