% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{geneID}
\alias{geneID}
\title{geneID generic}
\usage{
geneID(x)
}
\arguments{
\item{x}{enrichResult object}
}
\value{
'geneID' return the 'geneID' column of the enriched result which can be converted to data.frame via 'as.data.frame'
}
\description{
geneID generic
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
geneID(x)
}
