% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{geom_gsea_gene}
\alias{geom_gsea_gene}
\title{geom_gsea_gene}
\usage{
geom_gsea_gene(
  genes,
  mapping = NULL,
  geom = ggplot2::geom_text,
  ...,
  geneSet = NULL
)
}
\arguments{
\item{genes}{selected genes to be labeled}

\item{mapping}{aesthetic mapping, default is NULL}

\item{geom}{geometric layer to plot the gene labels, default is geom_text}

\item{...}{additional parameters passed to the 'geom'}

\item{geneSet}{choose which gene set(s) to be label if the plot contains multiple gene sets}
}
\value{
ggplot object
}
\description{
label genes in running score plot
}
\author{
Guangchuang Yu
}
