% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{get_enrichplot_color}
\alias{get_enrichplot_color}
\title{Color utility functions for enrichplot package}
\usage{
get_enrichplot_color(n = 2)
}
\arguments{
\item{n}{number of colors (2 or 3)}
}
\value{
color vector
}
\description{
This file contains all color-related helper functions
Get default enrichplot colors
}
