% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emapplot.R
\name{get_igraph}
\alias{get_igraph}
\title{Get an iGraph object}
\usage{
get_igraph(x, y, n, color, line_scale, min_edge, method, semData)
}
\arguments{
\item{x}{enrichment result.}

\item{y}{as.data.frame(x).}

\item{n}{number of enriched terms to display.}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue.}

\item{line_scale}{scale of line width.}

\item{min_edge}{minimum percentage of overlap genes to display the edge, should between 0 and 1, default value is 0.2.}

\item{method}{method of calculating the similarity between nodes, one of "Resnik", 
"Lin", "Rel", "Jiang" , "Wang"  and "JC"(Jaccard similarity coefficient) methods}

\item{semData}{GOSemSimDATA object}
}
\value{
an iGraph object
}
\description{
Get an iGraph object
}
