% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtable.R
\name{ggtable}
\alias{ggtable}
\title{ggtable}
\usage{
ggtable(d, p = NULL)
}
\arguments{
\item{d}{data frame}

\item{p}{ggplot object to extract color to color rownames(d), optional}
}
\value{
ggplot object
}
\description{
plot table
}
\author{
guangchuang yu
}
