% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityplot.R
\name{gseadist}
\alias{gseadist}
\title{gseadist}
\usage{
gseadist(x, IDs, type = "density")
}
\arguments{
\item{x}{GSEA result}

\item{IDs}{gene set IDs}

\item{type}{one of 'density' or 'boxplot'}
}
\value{
distribution plot
}
\description{
plot logFC distribution of selected gene sets
}
\author{
Guangchuang Yu
}
