% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-AllClasses.R
\docType{class}
\name{gseaResult-class}
\alias{gseaResult-class}
\alias{gseahResult-class}
\alias{show,gseaResult-method}
\title{Class "gseaResult"
This class represents the result of GSEA analysis}
\description{
Class "gseaResult"
This class represents the result of GSEA analysis
}
\section{Slots}{

\describe{
\item{\code{result}}{GSEA anaysis}

\item{\code{organism}}{organism}

\item{\code{setType}}{setType}

\item{\code{geneSets}}{geneSets}

\item{\code{geneList}}{order rank geneList}

\item{\code{keytype}}{ID type of gene}

\item{\code{permScores}}{permutation scores}

\item{\code{params}}{parameters}

\item{\code{gene2Symbol}}{gene ID to Symbol}

\item{\code{readable}}{whether convert gene ID to symbol}
}}

\seealso{
\code{\link{gseaplot}}
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}
}
\keyword{classes}
