% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gsInfo}
\alias{gsInfo}
\title{gsInfo}
\usage{
gsInfo(object, geneSetID, ...)
}
\arguments{
\item{object}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{...}{additional arguments passed to methods (e.g. \code{layer} for \code{mnseaResult})}
}
\value{
data.frame
}
\description{
extract gsea result of selected geneSet
}
\author{
Guangchuang Yu
}
