% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emapplot.R
\name{has_pairsim}
\alias{has_pairsim}
\title{Check whether the similarity matrix exists}
\usage{
has_pairsim(x)
}
\arguments{
\item{x}{result of enrichment analysis}
}
\description{
Check whether the similarity matrix exists
}
