% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/manhattanplot.R
\name{manhattanplot}
\alias{manhattanplot}
\alias{manhattanplot,enrichResult-method}
\alias{manhattanplot,gseaResult-method}
\alias{manhattanplot,compareClusterResult-method}
\alias{manhattanplot,compareClusterResult,ANY-method}
\alias{manhattanplot,enrichResultList-method}
\alias{manhattanplot,enrichResultList,ANY-method}
\alias{manhattanplot,gseaResultList-method}
\alias{manhattanplot,gseaResultList,ANY-method}
\alias{manhattanplot,list-method}
\alias{manhattanplot,list,ANY-method}
\title{manhattanplot}
\usage{
manhattanplot(x, ...)

\S4method{manhattanplot}{enrichResult}(
  x,
  color = "p.adjust",
  showCategory = 5,
  size = "Count",
  split = NULL,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)

\S4method{manhattanplot}{gseaResult}(
  x,
  color = "p.adjust",
  showCategory = 5,
  size = "Count",
  split = NULL,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)

\S4method{manhattanplot}{compareClusterResult}(
  x,
  color = "p.adjust",
  showCategory = 5,
  split = NULL,
  font.size = 12,
  title = "",
  size = "Count",
  includeAll = TRUE,
  label_format = 30,
  ...
)

\S4method{manhattanplot}{enrichResultList}(
  x,
  color = "p.adjust",
  showCategory = 5,
  size = "Count",
  split = NULL,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)

\S4method{manhattanplot}{gseaResultList}(
  x,
  color = "p.adjust",
  showCategory = 5,
  size = "Count",
  split = NULL,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)

\S4method{manhattanplot}{list}(
  x,
  color = "p.adjust",
  showCategory = 5,
  size = "Count",
  split = NULL,
  font.size = 12,
  title = "",
  label_format = 30,
  ...
)
}
\arguments{
\item{x}{enrichment result.}

\item{...}{additional parameters.}

\item{color}{variable used to color enriched terms, e.g. \code{'pvalue'},
\code{'p.adjust'}, or \code{'qvalue'}.}

\item{showCategory}{number of categories to display, or a vector of terms.}

\item{size}{variable used to scale category size, one of \code{"geneRatio"},
\code{"Percentage"}, or \code{"count"}.}

\item{split}{apply \code{showCategory} to each category specified by \code{split},
e.g., \code{"ONTOLOGY"}, \code{"category"}, or \code{"intersect"}. Default is \code{NULL}.}

\item{font.size}{font size.}

\item{title}{figure title.}

\item{label_format}{a numeric wrap width, or a custom function to format
axis labels.}

\item{includeAll}{logical value.}
}
\value{
ggplot object
}
\description{
Manhattan plot for enrichment result
}
\author{
Guangchuang Yu
}
