% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_utils.R
\name{plotting.clusterProfile}
\alias{plotting.clusterProfile}
\title{Internal plot function for plotting compareClusterResult}
\usage{
plotting.clusterProfile(
  clProf.reshape.df,
  x = ~Cluster,
  type = "dot",
  colorBy = "p.adjust",
  by = "geneRatio",
  title = "",
  font.size = 12,
  width = NULL
)
}
\arguments{
\item{clProf.reshape.df}{data frame of compareCluster result}

\item{x}{x variable}

\item{type}{one of dot and bar}

\item{colorBy}{one of pvalue or p.adjust}

\item{by}{one of percentage and count}

\item{title}{graph title}

\item{font.size}{graph font size}

\item{width}{width of the plot}
}
\value{
ggplot object
}
\description{
Internal plot function for plotting compareClusterResult
}
\author{
Guangchuang Yu \url{https://yulab-smu.top}
}
