% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/rewireplot.R
\name{rewireplot}
\alias{rewireplot}
\alias{rewireplot,nseaResult-method}
\alias{rewireplot,mnseaResult-method}
\title{rewireplot}
\usage{
rewireplot(x, ...)

\S4method{rewireplot}{nseaResult}(x, ...)

\S4method{rewireplot}{mnseaResult}(x, ...)
}
\arguments{
\item{x}{A \code{mnseaResult} object.}

\item{...}{Additional parameters passed to plot methods.}
}
\value{
A ggplot object.
}
\description{
Pathway-specific rewiring plot
}
