% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_utils.R
\name{set_enrichplot_color}
\alias{set_enrichplot_color}
\title{Helper function to set color scale for enrichplot}
\usage{
set_enrichplot_color(
  colors = get_enrichplot_color(2),
  type = "color",
  name = NULL,
  .fun = NULL,
  reverse = TRUE,
  transform = "identity",
  ...
)
}
\arguments{
\item{colors}{user provided color vector}

\item{type}{one of 'color', 'colour' or 'fill'}

\item{name}{name of the color legend}

\item{.fun}{force to use user provided color scale function}

\item{reverse}{whether reverse the color scheme}

\item{transform}{transform the color scale}

\item{...}{additional parameters}
}
\value{
a color scale
}
\description{
Helper function to set color scale for enrichplot
}
\author{
Guangchuang Yu
}
