% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show-method.R
\docType{methods}
\name{show}
\alias{show}
\alias{show}
\title{show method}
\usage{
show(object)

show(object)
}
\arguments{
\item{object}{A \code{enrichResult} instance.}
}
\value{
message

message
}
\description{
show method for \code{enrichResult} instance

show method for \code{gseaResult} instance
}
\author{
Guangchuang Yu \url{https://guangchuangyu.github.io}

Guangchuang Yu \url{https://guangchuangyu.github.io}
}
