% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/upsetplot.R
\docType{methods}
\name{upsetplot}
\alias{upsetplot}
\alias{upsetplot,enrichResult-method}
\alias{upsetplot,enrichResult,ANY-method}
\alias{upsetplot,gseaResult-method}
\alias{upsetplot,gseaResult}
\alias{upsetplot.enrichResult}
\alias{upsetplot.gseaResult}
\title{upsetplot method}
\usage{
upsetplot(x, ...)

\S4method{upsetplot}{enrichResult}(x, n = 10, ...)

\S4method{upsetplot}{gseaResult}(x, n = 10, ...)

upsetplot.enrichResult(x, n = 10, enrichment_only = TRUE, ...)

upsetplot.gseaResult(
  x,
  n = 10,
  type = "boxplot",
  nodeSize = 2,
  enrichment_only = TRUE,
  ...
)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters}

\item{n}{A number or a list of terms. If it is a number, 
the first n terms will be displayed. If it is a list of terms, 
the selected terms will be displayed.}

\item{enrichment_only}{whether only using enriched gene(for enrichResult) or core_enriched genes(for gseaResult)}

\item{type}{one of "boxplot" and "geom_violin"}

\item{nodeSize}{Size of nodes}
}
\value{
plot
}
\description{
upsetplot method generics
}
\examples{
require(DOSE)
data(geneList)
de=names(geneList)[1:100]
x <- enrichDO(de)
upsetplot(x, 8)
}
\author{
Guangchuang Yu
}
