% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud.r
\name{wordcloud_i}
\alias{wordcloud_i}
\title{Get cluster labels}
\usage{
wordcloud_i(cluster, pdata2, N)
}
\arguments{
\item{cluster}{a cluster name}

\item{pdata2}{the data section of the ggraph object}

\item{N}{the number of words in the cluster tags}
}
\description{
Get cluster labels
}
