% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{autofacet}
\alias{autofacet}
\title{automatically split barplot or dotplot into several facets}
\usage{
autofacet(by = "row", scales = "free", levels = NULL)
}
\arguments{
\item{by}{one of 'row' or 'column'}

\item{scales}{wether 'fixed' or 'free'}

\item{levels}{set facet levels}
}
\value{
a ggplot object
}
\description{
automatically split barplot or dotplot into several facets
}
