% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot.enrichResult}
\alias{barplot.enrichResult}
\title{barplot}
\usage{
\method{barplot}{enrichResult}(
  height,
  x = "Count",
  color = "p.adjust",
  showCategory = 8,
  font.size = 12,
  title = "",
  ...
)
}
\arguments{
\item{height}{enrichResult object}

\item{x}{one of 'Count' and 'GeneRatio'}

\item{color}{one of 'pvalue', 'p.adjust', 'qvalue'}

\item{showCategory}{number of categories to show}

\item{font.size}{font size}

\item{title}{plot title}

\item{...}{other parameter, ignored}
}
\value{
ggplot object
}
\description{
barplot of enrichResult
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
barplot(x)
}
