% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnetplot.R
\name{cnetplot}
\alias{cnetplot}
\title{cnetplot}
\usage{
cnetplot(x, showCategory = 5, foldChange = NULL, layout = "kk",
  circular = FALSE, ...)
}
\arguments{
\item{x}{enrichment result. e.g. instance of gseaResult or enrichResult}

\item{showCategory}{number of enriched terms to display}

\item{foldChange}{fold Change}

\item{layout}{layout of the network}

\item{circular}{whether using circular layout}

\item{...}{additional parameters}
}
\value{
ggplot object
}
\description{
gene-concept network
}
\author{
guangchuang yu
}
