% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cnetplot.R
\name{cnetplot}
\alias{cnetplot}
\alias{cnetplot,enrichResult-method}
\alias{cnetplot,list-method}
\alias{cnetplot,gseaResult-method}
\alias{cnetplot,compareClusterResult-method}
\alias{cnetplot.enrichResult}
\title{cnetplot}
\usage{
cnetplot(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{enrichResult}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{list}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{gseaResult}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{compareClusterResult}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

cnetplot.enrichResult(
  x,
  showCategory = 5,
  foldChange = NULL,
  layout = "kk",
  colorEdge = FALSE,
  circular = FALSE,
  node_label = "all",
  cex_category = 1,
  cex_gene = 1,
  cex_label_category = 1,
  cex_label_gene = 1,
  color_category = "#E5C494",
  color_gene = "#B3B3B3",
  shadowtext = "all",
  hilight.params = list(category = NULL, alpha_hilight = 1, alpha_no_hilight = 0.3),
  ...
)
}
\arguments{
\item{x}{Enrichment result.}

\item{showCategory}{A number or a vector of terms. If it is a number, 
the first n terms will be displayed. If it is a vector of terms, 
the selected terms will be displayed.}

\item{foldChange}{Fold Change of nodes, the default value is NULL. 
If the user provides the Fold Change value of the nodes, 
it can be used to set the color of the gene node.}

\item{layout}{Layout of the map, e.g. 'star', 'circle', 'gem', 'dh', 'graphopt', 'grid', 'mds', 
'randomly', 'fr', 'kk', 'drl' or 'lgl'.}

\item{...}{Additional parameters}

\item{colorEdge}{Logical, whether coloring edge by enriched terms, the default value is FALSE.}

\item{circular}{Logical, whether using circular layout, the default value is FALSE.}

\item{node_label}{Select which labels to be displayed.
one of 'category', 'gene', 'all'(the default) and 'none'.}

\item{cex_category}{Number indicating the amount by which plotting category
nodes should be scaled relative to the default, the default value is 1.}

\item{cex_gene}{Number indicating the amount by which plotting gene nodes
should be scaled relative to the default, the default value is 1.}

\item{cex_label_category}{Scale of category node label size, the 
default value is 1.}

\item{cex_label_gene}{Scale of gene node label size, the default value is 1.}

\item{color_category}{Color of category node.}

\item{color_gene}{Color of gene node.}

\item{shadowtext}{select which node labels to use shadow font,
one of 'category', 'gene', 'all' and 'none', default is 'all'.}

\item{hilight.params}{list, the parameters to control the attributes of highlighted nodes and edges.
see the hilight.params in the following.
hilight.params control the attributes of highlight, it can be referred to the following parameters:
    \itemize{
        \item \code{category} category nodes to be highlight.
        \item \code{alpha_hilight} alpha of highlighted nodes.
        \item \code{alpha_no_hilight} alpha of unhighlighted nodes.
    }}
}
\value{
ggplot object
}
\description{
Gene-Concept Network
}
\details{
plot linkages of genes and enriched concepts
(e.g. GO categories, KEGG pathways)
}
\examples{
\dontrun{
    library(DOSE)
    data(geneList)
    de <- names(geneList)[1:100]
    x <- enrichDO(de)
    x2 <- pairwise_termsim(x)
    cnetplot(x2)
    # use `layout` to change the layout of map
    cnetplot(x2, layout = "star")
    # use `showCategory` to select the displayed terms. It can be a number of a vector of terms.
    cnetplot(x2, showCategory = 10)
    categorys <- c("pre-malignant neoplasm", "intestinal disease",
                   "breast ductal carcinoma", "non-small cell lung carcinoma")
    cnetplot(x2, showCategory = categorys)
    # 'compareClusterResult' object is also supported.
    data(gcSample)
    library(clusterProfiler)
    library(DOSE)
    library(org.Hs.eg.db)
    data(gcSample)
    xx <- compareCluster(gcSample, fun="enrichGO", OrgDb="org.Hs.eg.db")
    xx2 <- pairwise_termsim(xx)
    cnetplot(xx2)
}
}
\author{
Guangchuang Yu
}
