% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cnetplot.R
\docType{methods}
\name{cnetplot}
\alias{cnetplot}
\alias{cnetplot,enrichResult-method}
\alias{cnetplot,gseaResult-method}
\alias{cnetplot.enrichResult}
\title{cnetplot}
\usage{
cnetplot(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{enrichResult}(x, showCategory = 5,
  foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{gseaResult}(x, showCategory = 5, foldChange = NULL,
  layout = "kk", ...)

cnetplot.enrichResult(x, showCategory = 5, foldChange = NULL,
  layout = "kk", colorEdge = FALSE, circular = FALSE,
  node_label = TRUE, ...)
}
\arguments{
\item{x}{enrichment result}

\item{showCategory}{number of enriched terms to display}

\item{foldChange}{fold Change}

\item{layout}{layout of the network}

\item{...}{additional parameters}

\item{colorEdge}{whether coloring edge by enriched terms}

\item{circular}{whether using circular layout}

\item{node_label}{whether display node label}
}
\value{
ggplot object
}
\description{
Gene-Concept Network
}
\details{
plot linkages of genes and enriched concepts
(e.g. GO categories, KEGG pathways)
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
cnetplot(x)
}
\author{
Guangchuang Yu
}
