% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/cnetplot.R
\name{cnetplot}
\alias{cnetplot}
\alias{cnetplot,enrichResult-method}
\alias{cnetplot,gseaResult-method}
\alias{cnetplot,compareClusterResult-method}
\alias{cnetplot.enrichResult}
\title{cnetplot}
\usage{
cnetplot(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{enrichResult}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{gseaResult}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

\S4method{cnetplot}{compareClusterResult}(x, showCategory = 5, foldChange = NULL, layout = "kk", ...)

cnetplot.enrichResult(
  x,
  showCategory = 5,
  foldChange = NULL,
  layout = "kk",
  colorEdge = FALSE,
  circular = FALSE,
  node_label = "all",
  cex_category = 1,
  cex_gene = 1,
  node_label_size = NULL,
  cex_label_category = 1,
  cex_label_gene = 1,
  ...
)
}
\arguments{
\item{x}{enrichment result}

\item{showCategory}{number of enriched terms to display}

\item{foldChange}{fold Change}

\item{layout}{layout of the network}

\item{...}{additional parameters}

\item{colorEdge}{whether coloring edge by enriched terms}

\item{circular}{whether using circular layout}

\item{node_label}{select which labels to be displayed.
one of 'category', 'gene', 'all' and 'none', default is "all".}

\item{cex_category}{number indicating the amount by which plotting category
nodes should be scaled relative to the default.}

\item{cex_gene}{number indicating the amount by which plotting gene nodes
should be scaled relative to the default.}

\item{node_label_size}{size of node label, this parameter has been
changed to cex_label_category and cex_label_gene}

\item{cex_label_category}{scale of category node label size}

\item{cex_label_gene}{scale of gene node label size}
}
\value{
ggplot object
}
\description{
Gene-Concept Network
}
\details{
plot linkages of genes and enriched concepts
(e.g. GO categories, KEGG pathways)
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
cnetplot(x)
}
\author{
Guangchuang Yu
}
