% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnetplot.R
\name{cnetplot.enrichResult}
\alias{cnetplot.enrichResult}
\alias{cnetplot.gseaResult}
\alias{cnetplot.compareClusterResult}
\title{cnetplot}
\usage{
\method{cnetplot}{enrichResult}(
  x,
  layout = igraph::layout_with_kk,
  showCategory = 5,
  color_category = "#E5C494",
  size_category = 1,
  color_item = "#B3B3B3",
  size_item = 1,
  color_edge = "grey",
  size_edge = 0.5,
  node_label = "all",
  foldChange = NULL,
  hilight = "none",
  hilight_alpha = 0.3,
  ...
)

\method{cnetplot}{gseaResult}(
  x,
  layout = igraph::layout_with_kk,
  showCategory = 5,
  color_category = "#E5C494",
  size_category = 1,
  color_item = "#B3B3B3",
  size_item = 1,
  color_edge = "grey",
  size_edge = 0.5,
  node_label = "all",
  foldChange = NULL,
  hilight = "none",
  hilight_alpha = 0.3,
  ...
)

\method{cnetplot}{compareClusterResult}(
  x,
  layout = igraph::layout_with_kk,
  showCategory = 5,
  color_category = "#E5C494",
  size_category = 1,
  color_item = "#B3B3B3",
  size_item = 1,
  color_edge = "grey",
  size_edge = 0.5,
  node_label = "all",
  foldChange = NULL,
  hilight = "none",
  hilight_alpha = 0.3,
  pie = "equal",
  ...
)
}
\arguments{
\item{x}{input object}

\item{layout}{network layout}

\item{showCategory}{selected category to be displayed}

\item{color_category}{color of category node}

\item{size_category}{relative size of the category}

\item{color_item}{color of item node}

\item{size_item}{relative size of the item (e.g., genes)}

\item{color_edge}{color of edge}

\item{size_edge}{relative size of edge}

\item{node_label}{one of 'all', 'none', 'category' and 'item'}

\item{foldChange}{numeric values to color the item (e.g, foldChange of gene expression values)}

\item{hilight}{selected category to be highlighted}

\item{hilight_alpha}{transparent value for not selected to be highlight}

\item{...}{additional parameters}

\item{pie}{one of 'equal' or 'Count' to set the slice ratio of the pies}
}
\description{
category-gene-network plot
}
\seealso{
[cnetplot][ggtangle::cnetplot]
}
