% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/consensusmap.R
\name{consensusmap}
\alias{consensusmap}
\alias{consensusmap,mnseaResult-method}
\alias{consensusmap,nseaResult-method}
\alias{consensusmap,list-method}
\title{consensusmap}
\usage{
consensusmap(x, ...)

\S4method{consensusmap}{mnseaResult}(x, ...)

\S4method{consensusmap}{nseaResult}(x, ...)

\S4method{consensusmap}{list}(x, ...)
}
\arguments{
\item{x}{A \code{mnseaResult}, \code{nseaResult}, or a named list of results.}

\item{...}{Additional parameters passed to plot methods.}
}
\value{
A ggplot object.
}
\description{
Multi-context mechanism consensus map
}
