% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot}
\alias{dotplot}
\alias{dotplot.enrichResult}
\alias{dotplot.gseaResult}
\title{dotplot}
\usage{
dotplot(height, x = "geneRatio", color = "p.adjust", showCategory = 10,
  split = NULL, font.size = 12, title = "")

\method{dotplot}{enrichResult}(height, x = "geneRatio", color = "p.adjust",
  showCategory = 10, split = NULL, font.size = 12, title = "")

\method{dotplot}{gseaResult}(height, x = "geneRatio", color = "p.adjust",
  showCategory = 10, split = NULL, font.size = 12, title = "")
}
\arguments{
\item{height}{input object, just name it to make it consistent with barplot}

\item{x}{variable for x-axis, one of 'geneRatio' or 'Count'}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{showCategory}{number of enriched terms to display}

\item{split}{separate result by 'category' variable}

\item{font.size}{font size}

\item{title}{plot title}
}
\value{
plot
}
\description{
dotplot for enrichment result
}
\author{
guangchuang yu

guangchuang yu
}
