% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/dotplot.R
\name{dotplot}
\alias{dotplot}
\alias{dotplot,enrichResult-method}
\alias{dotplot,gseaResult-method}
\title{dotplot}
\usage{
dotplot(object, ...)

\S4method{dotplot}{enrichResult}(
  object,
  x = "GeneRatio",
  color = "p.adjust",
  showCategory = 10,
  size = NULL,
  split = NULL,
  font.size = 12,
  title = "",
  ...
)

\S4method{dotplot}{gseaResult}(
  object,
  x = "GeneRatio",
  color = "p.adjust",
  showCategory = 10,
  size = NULL,
  split = NULL,
  font.size = 12,
  title = "",
  ...
)
}
\arguments{
\item{object}{input object}

\item{...}{additional parameters}

\item{x}{variable for x-axis, one of 'GeneRatio' or 'Count'}

\item{color}{variable that used to color enriched terms,
e.g. pvalue, p.adjust or qvalue}

\item{showCategory}{number of enriched terms to display}

\item{size}{variable that used to scale the sizes of categories}

\item{split}{separate result by 'category' variable}

\item{font.size}{font size}

\item{title}{plot title}
}
\value{
plot
}
\description{
dotplot for enrichment result
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
dotplot(x)
}
\author{
guangchuang yu
}
