% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/emapplot.R
\name{emapplot}
\alias{emapplot}
\alias{emapplot,enrichResult-method}
\alias{emapplot,gseaResult-method}
\alias{emapplot,compareClusterResult-method}
\alias{emapplot,mnseaResult-method}
\alias{prepare_emapplot_data}
\title{emapplot}
\usage{
emapplot(
  x,
  layout = igraph::layout_with_kk,
  coords = NULL,
  showCategory = 30,
  color = "p.adjust",
  size_category = 1,
  min_edge = 0.2,
  color_edge = "grey",
  size_edge = 0.5,
  group = NULL,
  group_legend = TRUE,
  node_label = "category",
  node_label_size = 5,
  pie = "equal",
  layer = NULL,
  label_format = 30,
  clusterFunction = stats::kmeans,
  nWords = 4,
  nCluster = NULL,
  ...
)

\S4method{emapplot}{enrichResult}(
  x,
  layout = igraph::layout_with_kk,
  coords = NULL,
  showCategory = 30,
  color = "p.adjust",
  size_category = 1,
  min_edge = 0.2,
  color_edge = "grey",
  size_edge = 0.5,
  group = NULL,
  group_legend = TRUE,
  node_label = "category",
  node_label_size = 5,
  pie = "equal",
  layer = NULL,
  label_format = 30,
  clusterFunction = stats::kmeans,
  nWords = 4,
  nCluster = NULL,
  show_category_size_legend = TRUE,
  ...
)

\S4method{emapplot}{gseaResult}(
  x,
  layout = igraph::layout_with_kk,
  coords = NULL,
  showCategory = 30,
  color = "p.adjust",
  size_category = 1,
  min_edge = 0.2,
  color_edge = "grey",
  size_edge = 0.5,
  group = NULL,
  group_legend = TRUE,
  node_label = "category",
  node_label_size = 5,
  pie = "equal",
  layer = NULL,
  label_format = 30,
  clusterFunction = stats::kmeans,
  nWords = 4,
  nCluster = NULL,
  show_category_size_legend = TRUE,
  ...
)

\S4method{emapplot}{compareClusterResult}(
  x,
  layout = igraph::layout_with_kk,
  coords = NULL,
  showCategory = 30,
  color = "p.adjust",
  size_category = 1,
  min_edge = 0.2,
  color_edge = "grey",
  size_edge = 0.5,
  group = NULL,
  group_legend = TRUE,
  node_label = "category",
  node_label_size = 5,
  pie = "equal",
  layer = NULL,
  label_format = 30,
  clusterFunction = stats::kmeans,
  nWords = 4,
  nCluster = NULL,
  show_category_size_legend = TRUE,
  ...
)

\S4method{emapplot}{mnseaResult}(
  x,
  layout = igraph::layout_with_kk,
  coords = NULL,
  showCategory = 30,
  color = "p.adjust",
  size_category = 1,
  min_edge = 0.2,
  color_edge = "grey",
  size_edge = 0.5,
  group = NULL,
  group_legend = TRUE,
  node_label = "category",
  node_label_size = 5,
  pie = "equal",
  layer = NULL,
  label_format = 30,
  clusterFunction = stats::kmeans,
  nWords = 4,
  nCluster = NULL,
  show_category_size_legend = TRUE,
  ...
)

prepare_emapplot_data(x, showCategory, color, min_edge, size_edge)
}
\arguments{
\item{x}{Enrichment result.}

\item{layout}{igraph layout}

\item{coords}{optional user-supplied coordinate data.frame with \code{x}, \code{y}
and row names matching node labels.}

\item{showCategory}{number of categories to display or a vector of terms.}

\item{color}{Variable used to color enriched terms, e.g. 'pvalue',
'p.adjust' or 'qvalue'.}

\item{size_category}{relative size of the categories}

\item{min_edge}{The minimum similarity threshold for whether
two nodes are connected, should be between 0 and 1, default value is 0.2.}

\item{color_edge}{color of the network edge}

\item{size_edge}{relative size of edge width}

\item{group}{logical, if TRUE, group the categories.}

\item{group_legend}{logical, if TRUE, draw a legend for the groups.}

\item{node_label}{Select which labels to display,
one of 'category', 'group', 'all' and 'none'.}

\item{node_label_size}{size of node label, default is 5.}

\item{pie}{one of 'equal' or 'Count' to set the slice ratio of the pies}

\item{layer}{optional layer or layers to retain for \code{mnseaResult} plots.}

\item{label_format}{a numeric value sets wrap length, alternatively a custom function to format axis labels.}

\item{clusterFunction}{clustering method function, such as \code{stats::kmeans} (default),
\code{cluster::clara}, \code{cluster::fanny}, or \code{cluster::pam}.}

\item{nWords}{Numeric, the number of words in the cluster tags, the default value is 4.}

\item{nCluster}{Numeric, the number of clusters,
the default value is square root of the number of nodes.}

\item{...}{Additional parameters}

\item{show_category_size_legend}{Logical, whether to draw the category-size
annotation legend for compareCluster pie charts.}
}
\value{
ggplot object
}
\description{
Enrichment Map for enrichment result of
over-representation test or gene set enrichment analysis
}
\details{
This function visualizes gene sets as a network (i.e. enrichment map).
Mutually overlapping gene sets tend to cluster together, making it
easier for interpretation. When the similarity between terms meets
a certain threshold (default is 0.2, adjusted by parameter \code{min_edge}),
there will be edges between terms. The stronger the similarity,
the shorter and thicker the edges. The similarity between terms is
obtained by the function \code{pairwise_termsim}. Details of the similarity
calculation can be found in its documentation: \code{\link[=pairwise_termsim]{pairwise_termsim()}}.
}
\examples{
\dontrun{
    library(DOSE)
    data(geneList)
    de <- names(geneList)[1:100]
    x <- enrichDO(de)
    x2 <- pairwise_termsim(x)
    emapplot(x2)
    # use `layout` to change the layout of map
    emapplot(x2, layout = "star")
    # use `showCategory` to  select the displayed terms. It can be a number of a vector of terms.
    emapplot(x2, showCategory = 10)
    categories <- c("pre-malignant neoplasm", "intestinal disease",
                   "breast ductal carcinoma")
    emapplot(x2, showCategory = categories)

    # It can also graph compareClusterResult
    library(clusterProfiler)
    library(DOSE)
    library(org.Hs.eg.db)
    data(gcSample)
    xx <- compareCluster(gcSample, fun="enrichGO", OrgDb="org.Hs.eg.db")
    xx2 <- pairwise_termsim(xx)
    emapplot(xx2)
}
}
\author{
Guangchuang Yu
}
