% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/emapplot.R
\name{emapplot}
\alias{emapplot}
\alias{emapplot,enrichResult-method}
\alias{emapplot,gseaResult-method}
\alias{emapplot,compareClusterResult-method}
\alias{emapplot.enrichResult}
\title{emapplot}
\usage{
emapplot(x, showCategory = 30, color = "p.adjust", layout = "kk", ...)

\S4method{emapplot}{enrichResult}(x, showCategory = 30, color = "p.adjust", layout = "nicely", ...)

\S4method{emapplot}{gseaResult}(x, showCategory = 30, color = "p.adjust", layout = "nicely", ...)

\S4method{emapplot}{compareClusterResult}(x, showCategory = 5, color = "p.adjust", layout = "nicely", ...)

emapplot.enrichResult(
  x,
  showCategory = 30,
  color = "p.adjust",
  layout = "nicely",
  pie_scale = 1,
  line_scale = 1,
  min_edge = 0.2,
  method = "other",
  measure = "Wang",
  OrgDb = NULL,
  ont = "BP",
  ...
)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{layout}{layout of the map}

\item{...}{additional parameters}

\item{pie_scale}{scale of pie plot}

\item{line_scale}{scale of line width}

\item{min_edge}{minimum percentage of overlap genes to display the edge, should between 0 and 1, default value is 0.2}

\item{method}{method of calculating the similarity between nodes, one of "goSim", "doSim" and "other"}

\item{measure}{method for calculating semantic similarity, one of "Resnik", "Lin", "Rel", "Jiang" and "Wang" methods}

\item{OrgDb}{OrgDb object}

\item{ont}{one of 'BP', 'MF', 'CC'}
}
\value{
ggplot object
}
\description{
Enrichment Map for enrichment result of
over-representation test or gene set enrichment analysis
}
\details{
This function visualizes gene sets as a network (i.e. enrichment map).
Mutually overlapping gene sets tend to cluster together, making it easier for interpretation.
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
emapplot(x)
}
\author{
Guangchuang Yu
}
