% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/emapplot.R
\name{emapplot}
\alias{emapplot}
\alias{emapplot.enrichResult}
\title{emapplot}
\usage{
emapplot(x, showCategory = 30, color = "p.adjust", layout = "kk", ...)

\method{emapplot}{enrichResult}(x, showCategory = 30, color = "p.adjust",
  layout = "kk", ...)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{layout}{layout of the map}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
Enrichment Map for enrichment result of
over-representation test or gene set enrichment analysis
}
\details{
This function visualizes gene sets as a network (i.e. enrichment map).
Mutually overlapping gene sets tend to cluster together, making it easier for interpretation.
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
emapplot(x)
}
\author{
Guangchuang Yu
}
