% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emapplot_cluster.r
\name{emapplot_cluster}
\alias{emapplot_cluster}
\title{Functional grouping network diagram for enrichment result of
over-representation test or gene set enrichment analysis}
\usage{
emapplot_cluster(
  x,
  showCategory = nrow(x),
  color = "p.adjust",
  line_scale = 0.1,
  with_edge = TRUE,
  method = "other",
  nWords = 4,
  nCluster = NULL,
  OrgDb = NULL,
  split = NULL,
  ont = "BP",
  measure = "Wang",
  min_edge = 0.2,
  cluster_label_scale = 1
)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{line_scale}{scale of line width}

\item{with_edge}{if TRUE, draw the edges of the network diagram}

\item{method}{method of calculating the similarity between nodes, one of "goSim", "doSim" and "other"}

\item{nWords}{the number of words in the cluster tags}

\item{nCluster}{the number of clusters}

\item{OrgDb}{OrgDb object}

\item{split}{separate result by 'category' variable}

\item{ont}{one of 'BP', 'MF', 'CC'}

\item{measure}{method for calculating semantic similarity, one of "Resnik", "Lin", "Rel", "Jiang" and "Wang" methods}

\item{min_edge}{minimum percentage of overlap genes to display the edge, should between 0 and 1, default value is 0.2}

\item{cluster_label_scale}{scale of cluster labels size}
}
\description{
Functional grouping network diagram for enrichment result of
over-representation test or gene set enrichment analysis
}
\examples{
\dontrun{
    library(clusterProfiler)
    data(geneList, package = "DOSE")
       gene <- names(geneList)[abs(geneList) > 2]
       ego <- enrichGO(gene  = gene,
        universe      = names(geneList),
        OrgDb         = org.Hs.eg.db,
        ont           = "CC",
        pAdjustMethod = "BH",
        pvalueCutoff  = 0.01,
        qvalueCutoff  = 0.05,
        readable      = TRUE)
     emapplot_cluster(ego)
   }
}
