% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emapplot_cluster.R
\name{emapplot_cluster}
\alias{emapplot_cluster}
\title{Functional grouping network diagram for enrichment result of
over-representation test or gene set enrichment analysis}
\usage{
emapplot_cluster(x, ...)
}
\arguments{
\item{x}{enrichment result}

\item{...}{additional parameters. Please refer to: \link{emapplot}.}
}
\value{
ggplot2 object
}
\description{
This function has been replaced by `emapplot`.
}
