% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichplot-package.R
\docType{package}
\name{enrichplot-package}
\alias{enrichplot}
\alias{enrichplot-package}
\title{enrichplot: Visualization of Functional Enrichment Result}
\description{
The 'enrichplot' package implements several visualization methods for interpreting functional enrichment results obtained from ORA or GSEA analysis. It is mainly designed to work with the 'clusterProfiler' package suite. All the visualization methods are developed based on 'ggplot2' graphics.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yulab-smu.top/biomedical-knowledge-mining-book/}
  \item Report bugs at \url{https://github.com/GuangchuangYu/enrichplot/issues}
}

}
\author{
\strong{Maintainer}: Guangchuang Yu \email{guangchuangyu@gmail.com} (\href{https://orcid.org/0000-0002-6485-8781}{ORCID})

Other contributors:
\itemize{
  \item Chun-Hui Gao \email{gaospecial@gmail.com} (\href{https://orcid.org/0000-0002-1445-7939}{ORCID}) [contributor]
}

}
\keyword{internal}
