% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method-fortify.R
\name{fortify.compareClusterResult}
\alias{fortify.compareClusterResult}
\alias{fortify.enrichResult}
\title{fortify}
\usage{
\method{fortify}{compareClusterResult}(
  model,
  data,
  showCategory = 5,
  by = "geneRatio",
  split = NULL,
  includeAll = TRUE,
  ...
)

\method{fortify}{enrichResult}(
  model,
  data,
  showCategory = 5,
  by = "Count",
  order = FALSE,
  drop = FALSE,
  split = NULL,
  ...
)
}
\arguments{
\item{model}{'enrichResult' or 'compareClusterResult' object}

\item{data}{not use here}

\item{showCategory}{Category numbers to show}

\item{by}{one of Count and GeneRatio}

\item{split}{separate result by 'split' variable}

\item{includeAll}{logical}

\item{...}{additional parameter}

\item{order}{logical}

\item{drop}{logical}
}
\value{
data.frame

data.frame
}
\description{
convert compareClusterResult to a data.frame that ready for plot

convert enrichResult object for ggplot2
}
\author{
Guangchuang Yu
}
