% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emapplot.R
\name{get_igraph}
\alias{get_igraph}
\title{Get an iGraph object}
\usage{
get_igraph(x, y, n, color, cex_line, min_edge)
}
\arguments{
\item{x}{enrichment result.}

\item{y}{as.data.frame(x).}

\item{n}{number of enriched terms to display.}

\item{color}{variable that used to color enriched terms, e.g. pvalue,
p.adjust or qvalue.}

\item{cex_line}{scale of line width.}

\item{min_edge}{minimum percentage of overlap genes to display the edge,
should between 0 and 1, default value is 0.2.}
}
\value{
an iGraph object
}
\description{
Get an iGraph object
}
