% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goplot.R
\name{goplot}
\alias{goplot}
\title{goplot}
\usage{
goplot(x, showCategory = 10, color = "p.adjust", layout = "sugiyama", ...)
}
\arguments{
\item{x}{enrichment result. e.g. instance of gseaResult or enrichResult}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{layout}{layout of the map}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
plot induced GO DAG of significant terms
}
\author{
guangchuang yu
}
