% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot}
\alias{gseaplot}
\title{visualize analyzing result of GSEA}
\usage{
gseaplot(gseaResult, geneSetID, by = "all", title = "", color = "black",
  color.line = "green", color.vline = "#FA5860")
}
\arguments{
\item{gseaResult}{gseaResult object}

\item{geneSetID}{geneSet ID}

\item{by}{one of "runningScore" or "position"}

\item{title}{plot title}

\item{color}{color of line segments}

\item{color.line}{color of running enrichment score line}

\item{color.vline}{color of vertical line which indicating the maximum/minimal running enrichment score}
}
\value{
ggplot2 object
}
\description{
plotting function for gseaResult
}
\examples{
library(DOSE)
data(geneList)
x <- gseDO(geneList)
gseaplot(x, geneSetID=1)
}
\author{
Yu Guangchuang
}
