% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot2}
\alias{gseaplot2}
\title{gseaplot2}
\usage{
gseaplot2(
  x,
  geneSetID,
  title = "",
  color = "green",
  base_size = 11,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  pvalue_table = FALSE,
  pvalue_table_columns = c("pvalue", "p.adjust"),
  pvalue_table_rownames = "Description",
  ES_geom = "line",
  layer = NULL
)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{title}{plot title}

\item{color}{color of running enrichment score line}

\item{base_size}{base font size}

\item{rel_heights}{relative heights of subplots}

\item{subplots}{which subplots to be displayed}

\item{pvalue_table}{whether add pvalue table}

\item{pvalue_table_columns}{selected columns to be plotted in the \code{pvalue_table}}

\item{pvalue_table_rownames}{selected column as the rownames of the \code{pvalue_table}. If set to NULL, no rownames will be displayed.}

\item{ES_geom}{geom for plotting running enrichment score,
one of 'line' or 'dot'}

\item{layer}{Optional \code{mnsea} layer. When \code{NULL}, use collapsed scores.}
}
\value{
plot
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
\author{
Guangchuang Yu
}
