% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot2}
\alias{gseaplot2}
\title{gseaplot2}
\usage{
gseaplot2(x, geneSetID, title = "", color.line = "green",
  color.vline = "#FA5860")
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{title}{plot title}

\item{color.line}{color of running enrichment score line}

\item{color.vline}{color of vertical line which indicating the maximum/minimal running enrichment score}
}
\value{
plot
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
\author{
guangchuang yu
}
