% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot2}
\alias{gseaplot2}
\title{gseaplot2}
\usage{
gseaplot2(x, geneSetID, title = "", color = "green", rel_heights = c(1.5,
  0.5, 1), subplots = 1:3, pvalue_table = FALSE)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{title}{plot title}

\item{color}{color of running enrichment score line}

\item{rel_heights}{relative heights of subplots}

\item{subplots}{which subplots to be displayed}

\item{pvalue_table}{whether add pvalue table}
}
\value{
plot
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
\author{
Guangchuang Yu
}
