% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gseaplot2}
\alias{gseaplot2}
\title{gseaplot2}
\usage{
gseaplot2(
  x,
  geneSetID,
  title = "",
  color = "green",
  base_size = 11,
  rel_heights = c(1.5, 0.5, 1),
  subplots = 1:3,
  pvalue_table = FALSE,
  ES_geom = "line",
  lineSize1 = 1,
  lineSize2 = 0.5,
  legendKeySize = 1
)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{title}{plot title}

\item{color}{color of running enrichment score line}

\item{base_size}{base font size}

\item{rel_heights}{relative heights of subplots}

\item{subplots}{which subplots to be displayed}

\item{pvalue_table}{whether add pvalue table}

\item{ES_geom}{geom for plotting running enrichment score,
one of 'line' or 'dot'}

\item{lineSize1}{Size of lines in enrichment score line chart}

\item{lineSize2}{Size of lines in the leading-edge subset.}

\item{legendKeySize}{Size of legend key in enrichment score line chart.}
}
\value{
plot
}
\description{
GSEA plot that mimic the plot generated by broad institute's GSEA software
}
\author{
Guangchuang Yu
}
