% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{gsearank}
\alias{gsearank}
\title{gsearank}
\usage{
gsearank(x, geneSetID, title = "", output = "plot", layer = NULL)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}

\item{title}{plot title}

\item{output}{one of 'plot' or 'table' (for exporting data)}
}
\value{
ggplot object
}
\description{
plot ranked list of genes with running enrichment score as bar height
}
\author{
Guangchuang Yu
}
