% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/heatplot.R
\name{heatplot}
\alias{heatplot}
\alias{heatplot,enrichResult-method}
\alias{heatplot,gseaResult-method}
\alias{heatplot,mnseaResult-method}
\alias{prepare_heatplot_data}
\title{heatplot}
\usage{
heatplot(x, showCategory = 30, ...)

\S4method{heatplot}{enrichResult}(x, showCategory = 30, ...)

\S4method{heatplot}{gseaResult}(x, showCategory = 30, ...)

\S4method{heatplot}{mnseaResult}(x, showCategory = 30, ...)

prepare_heatplot_data(x, showCategory, showTop, foldChange, pvalue)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{...}{Additional parameters}

\item{showTop}{number of top genes ranked by \code{abs(foldChange) * frequency}
to be shown in the heatmap, default NULL means all genes are shown}

\item{foldChange}{fold change.}

\item{pvalue}{pvalue of genes}

\item{label_format}{a numeric value sets wrap length, alternatively a
custom function to format axis labels.
by default wraps names longer than 30 characters}

\item{symbol}{symbol of the nodes, one of "rect" (the default) or "dot"}

\item{pathway_id}{optional pathway ID for pathway-specific \code{mnseaResult}
heatmaps.}

\item{layer}{optional layer or layers to retain for \code{mnseaResult} plots.}

\item{value}{fill value for \code{mnseaResult}; use \code{"share"} or
\code{"contribution"} for term-layer heatmaps and \code{"score"} or \code{"abs_score"}
for pathway-specific feature heatmaps.}
}
\value{
ggplot object
}
\description{
Heatmap-like plot for functional classification
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
heatplot(x)
}
\author{
Guangchuang Yu
}
