% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/heatplot.R
\name{heatplot}
\alias{heatplot}
\alias{heatplot,enrichResult-method}
\alias{heatplot,gseaResult-method}
\alias{heatplot.enrichResult}
\title{heatplot}
\usage{
heatplot(x, showCategory = 30, ...)

\S4method{heatplot}{enrichResult}(x, showCategory = 30, ...)

\S4method{heatplot}{gseaResult}(x, showCategory = 30, ...)

heatplot.enrichResult(
  x,
  showCategory = 30,
  showTop = NULL,
  symbol = "rect",
  foldChange = NULL,
  pvalue = NULL,
  label_format = 30
)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{...}{Additional parameters}

\item{showTop}{number of top genes ranked by \code{abs(foldChange) * frequency}
to be shown in the heatmap, default NULL means all genes are shown}

\item{symbol}{symbol of the nodes, one of "rect" (the default) or "dot"}

\item{foldChange}{fold change.}

\item{pvalue}{pvalue of genes}

\item{label_format}{a numeric value sets wrap length, alternatively a
custom function to format axis labels.
by default wraps names longer than 30 characters}
}
\value{
ggplot object
}
\description{
Heatmap-like plot for functional classification
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
heatplot(x)
}
\author{
Guangchuang Yu
}
