% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.R
\name{heatplot}
\alias{heatplot}
\title{heatplot}
\usage{
heatplot(x, showCategory = 30, foldChange = NULL)
}
\arguments{
\item{x}{enrichment result. e.g. instance of gseaResult or enrichResult}

\item{showCategory}{number of enriched terms to display}

\item{foldChange}{fold Change}
}
\value{
ggplot object
}
\description{
heatmap like plot for functional classification
}
\author{
guangchuang yu
}
