% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gseaplot.R
\name{hplot}
\alias{hplot}
\title{hplot}
\usage{
hplot(x, geneSetID)
}
\arguments{
\item{x}{gseaResult object}

\item{geneSetID}{gene set ID}
}
\value{
horizontal plot
}
\description{
Horizontal plot for GSEA result
}
\author{
Guangchuang Yu
}
