% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapplot.R
\name{mapplot}
\alias{mapplot}
\title{mapplot}
\usage{
mapplot(x, showCategory = 50, color = "p.adjust", layout = "kk", ...)
}
\arguments{
\item{x}{enrichment result. e.g. instance of gseaResult or enrichResult}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue, p.adjust or qvalue}

\item{layout}{layout of the map}

\item{...}{additional parameter}
}
\value{
ggplot object
}
\description{
Enrichment Map
}
\author{
Guangchuang Yu
}
