% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/mechanismflow.R
\name{mechanismflow}
\alias{mechanismflow}
\alias{mechanismflow,mnseaResult-method}
\alias{mechanismflow,nseaResult-method}
\alias{mechanismflow,list-method}
\title{mechanismflow}
\usage{
mechanismflow(x, ...)

\S4method{mechanismflow}{mnseaResult}(x, ...)

\S4method{mechanismflow}{nseaResult}(x, ...)

\S4method{mechanismflow}{list}(x, ...)
}
\arguments{
\item{x}{A \code{mnseaResult}, \code{nseaResult}, or a named list of results.}

\item{...}{Additional parameters passed to plot methods.}
}
\value{
A ggplot object.
}
\description{
Pathway state transition flow
}
