% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/phaseplot.R
\name{phaseplot}
\alias{phaseplot}
\alias{phaseplot,nseaResult-method}
\alias{phaseplot,mnseaResult-method}
\title{phaseplot}
\usage{
phaseplot(x, ...)

\S4method{phaseplot}{nseaResult}(x, ...)

\S4method{phaseplot}{mnseaResult}(x, ...)
}
\arguments{
\item{x}{A \code{nseaResult} or \code{mnseaResult} object.}

\item{...}{Additional parameters passed to plot methods.}
}
\value{
A ggplot object.
}
\description{
Phase plot for enrichment-shift versus rewiring
}
