% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmcplot.R
\name{pmcplot}
\alias{pmcplot}
\title{pmcplot}
\usage{
pmcplot(query, period, proportion = TRUE)
}
\arguments{
\item{query}{query terms}

\item{period}{period of query in the unit of year}

\item{proportion}{If TRUE, use query_hits/all_hits, otherwise use query_hits}
}
\value{
ggplot object
}
\description{
PubMed Central Trend plot
}
\author{
guangchuang yu
}
