% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggtitle}
\alias{facet_grid}
\alias{plot_list}
\alias{cnetplot}
\alias{theme_dose}
\alias{geom_cnet_label}
\alias{gseaScores}
\alias{geneID}
\alias{geneInCategory}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{aplot}{\code{\link[aplot:plot_list]{plot_list()}}}

  \item{DOSE}{\code{\link[DOSE:theme_dose]{theme_dose()}}}

  \item{enrichit}{\code{\link[enrichit:geneID]{geneID()}}, \code{\link[enrichit:geneInCategory]{geneInCategory()}}, \code{\link[enrichit:gseaScores]{gseaScores()}}}

  \item{ggplot2}{\code{\link[ggplot2:facet_grid]{facet_grid()}}, \code{\link[ggplot2:ggtitle]{ggtitle()}}}

  \item{ggtangle}{\code{\link[ggtangle:cnetplot]{cnetplot()}}, \code{\link[ggtangle:geom_cnet_label]{geom_cnet_label()}}}
}}

