% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{theme_dose}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{DOSE}{\code{\link[DOSE]{theme_dose}}}
}}

