% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggtitle}
\alias{plot_grid}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{cowplot}{\code{\link[cowplot]{plot_grid}}}

  \item{ggplot2}{\code{\link[ggplot2]{ggtitle}}}
}}

