% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{ggtitle}
\alias{facet_grid}
\alias{plot_list}
\alias{cnetplot}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{aplot}{\code{\link[aplot]{plot_list}}}

  \item{ggplot2}{\code{\link[ggplot2]{facet_grid}}, \code{\link[ggplot2:labs]{ggtitle}}}

  \item{ggtangle}{\code{\link[ggtangle]{cnetplot}}}
}}

