% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgeplot.R
\name{ridgeplot}
\alias{ridgeplot}
\title{ridgeplot}
\usage{
ridgeplot(x, showCategory = 30, fill = "p.adjust", core_enrichment = TRUE)
}
\arguments{
\item{x}{gseaResult object}

\item{showCategory}{number of categories for plotting}

\item{fill}{one of "pvalue", "p.adjust", "qvalue"}

\item{core_enrichment}{whether only using core_enriched genes}
}
\value{
ggplot object
}
\description{
ridgeline plot for GSEA result
}
\examples{
library(DOSE)
data(geneList)
x <- gseDO(geneList)
ridgeplot(x)
}
\author{
guangchuang yu
}
