% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_enrichplot_color}
\alias{set_enrichplot_color}
\title{set_enrichplot_color}
\usage{
set_enrichplot_color(
  colors = get_enrichplot_color(2),
  type = "color",
  name = NULL,
  .fun = NULL,
  reverse = TRUE,
  ...
)
}
\arguments{
\item{colors}{user provided color vector}

\item{type}{one of 'color', 'colour' or 'fill'}

\item{name}{name of the color legend}

\item{.fun}{force to use user provided color scale function}

\item{reverse}{whether reverse the color scheme, default is TRUE as it is more significant for lower pvalue}

\item{...}{additional parameter that passed to the color scale function}
}
\value{
a color scale
}
\description{
helper function to set color for enrichplot
}
