% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ssplot.R
\name{ssplot}
\alias{ssplot}
\alias{ssplot,enrichResult-method}
\alias{ssplot,gseaResult-method}
\alias{ssplot,compareClusterResult-method}
\alias{ssplot.enrichResult}
\alias{ssplot.compareClusterResult}
\title{ssplot}
\usage{
ssplot(x, showCategory = 30, ..., min_edge = 0.2, size_edge = 0.5)

\S4method{ssplot}{enrichResult}(x, showCategory = 30, ..., min_edge = 0.2, size_edge = 0.5)

\S4method{ssplot}{gseaResult}(x, showCategory = 30, ..., min_edge = 0.2, size_edge = 0.5)

\S4method{ssplot}{compareClusterResult}(x, showCategory = 30, ..., min_edge = 0.2, size_edge = 0.5)

ssplot.enrichResult(
  x,
  showCategory = 30,
  drfun = NULL,
  dr.params = list(),
  node_label = "group",
  min_edge = 0.2,
  size_edge = 0.5,
  ...
)

ssplot.compareClusterResult(
  x,
  showCategory = 30,
  pie = "equal",
  drfun = NULL,
  dr.params = list(),
  node_label = "group",
  min_edge = 0.2,
  size_edge = 0.5,
  ...
)
}
\arguments{
\item{x}{Enrichment result.}

\item{...}{additional parameters

Additional plotting parameters are inherited from \code{\link[=emapplot]{emapplot()}}.}

\item{showCategory}{number of categories to display or a vector of terms.}

\item{drfun}{Function used for dimension reduction. The default, \code{stats::cmdscale}, performs classical multidimensional scaling (MDS) and is one available reduction method.}

\item{dr.params}{Named list of arguments passed separately to \code{drfun} via \code{tidydr::dr}.}

\item{min_edge}{Minimum similarity threshold for connecting two nodes.}

\item{size_edge}{Relative size of edge width.}

\item{node_label}{Select which labels to display,
one of 'category', 'group', 'all' and 'none'.}

\item{pie}{one of 'equal' or 'Count' to set the slice ratio of the pies}
}
\value{
ggplot object
}
\description{
Similarity Space Plot for enrichment analysis
}
\details{
Creates 2D visualization of enrichment results using dimension reduction
techniques to show relationships between terms based on similarity.
}
\examples{
\dontrun{
    library(clusterProfiler)
    library(org.Hs.eg.db)
    library(enrichplot)
    library(GOSemSim)
    library(DOSE)
    data(geneList)
    gene <- names(geneList)[abs(geneList) > 2]
    ego <- enrichGO(gene  = gene,
        universe      = names(geneList),
        OrgDb         = org.Hs.eg.db,
        ont           = "BP",
        pAdjustMethod = "BH",
        pvalueCutoff  = 0.01,
        qvalueCutoff  = 0.05,
        readable      = TRUE)
    d <- godata('org.Hs.eg.db', ont="BP")
    ego2 <- pairwise_termsim(ego, method = "Wang", semData = d)
    ssplot(ego2)
}
}
\author{
Guangchuang Yu
}
