% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ssplot.R
\name{ssplot}
\alias{ssplot}
\alias{ssplot,enrichResult-method}
\alias{ssplot,gseaResult-method}
\alias{ssplot,compareClusterResult-method}
\alias{ssplot.enrichResult}
\alias{ssplot.compareClusterResult}
\title{ssplot}
\usage{
ssplot(x, ...)

\S4method{ssplot}{enrichResult}(x, showCategory = 30, ...)

\S4method{ssplot}{gseaResult}(x, showCategory = 30, ...)

\S4method{ssplot}{compareClusterResult}(x, showCategory = 30, ...)

ssplot.enrichResult(
  x,
  showCategory = 30,
  nCluster = NULL,
  drfun = stats::cmdscale,
  with_edge = FALSE,
  ...
)

ssplot.compareClusterResult(
  x,
  showCategory = 30,
  split = NULL,
  pie = "equal",
  nCluster = NULL,
  drfun = stats::cmdscale,
  with_edge = FALSE,
  cex_pie2axis = 0.0125,
  ...
)
}
\arguments{
\item{x}{enrichment result.}

\item{...}{additional parameters

additional parameters can refer the emapplot function: \link{emapplot}.}

\item{showCategory}{number of enriched terms to display}

\item{nCluster}{Numeric, the number of clusters,
the default value is square root of the number of nodes.}

\item{drfun}{The function used for dimension reduction,
e.g. stats::cmdscale (the default), vegan::metaMDS, or ape::pcoa.}

\item{with_edge}{Logical, if TRUE (the default), draw the edges of the network diagram.}

\item{split}{separate result by 'category' variable}

\item{pie}{proportion of clusters in the pie chart, one of 'equal' (default) and 'Count'}

\item{cex_pie2axis}{It is used to adjust the relative size of the pie chart on the coordinate axis,
the default value is 0.0125.}
}
\value{
ggplot object
}
\description{
Similarity space plot of enrichment analysis results.
}
\examples{
\dontrun{
    library(clusterProfiler)
    library(org.Hs.eg.db)
    library(enrichplot)
    library(GOSemSim)
    library(DOSE)
    data(geneList)
    gene <- names(geneList)[abs(geneList) > 2]
    ego <- enrichGO(gene  = gene,
        universe      = names(geneList),
        OrgDb         = org.Hs.eg.db,
        ont           = "BP",
        pAdjustMethod = "BH",
        pvalueCutoff  = 0.01,
        qvalueCutoff  = 0.05,
        readable      = TRUE)
    d <- godata('org.Hs.eg.db', ont="BP")
    ego2 <- pairwise_termsim(ego, method = "Wang", semData = d)
    ssplot(ego2)    
}
}
