% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/treeplot.r
\name{treeplot}
\alias{treeplot}
\alias{treeplot,enrichResult-method}
\alias{treeplot,gseaResult-method}
\alias{treeplot,compareClusterResult-method}
\alias{treeplot.enrichResult}
\alias{treeplot.compareClusterResult}
\title{treeplot}
\usage{
treeplot(x, showCategory = 30, color = "p.adjust", label_format = 30, ...)

\S4method{treeplot}{enrichResult}(x, showCategory = 30, color = "p.adjust", label_format = 30, ...)

\S4method{treeplot}{gseaResult}(x, showCategory = 30, color = "p.adjust", label_format = 30, ...)

\S4method{treeplot}{compareClusterResult}(x, showCategory = 5, color = "p.adjust", label_format = 30, ...)

treeplot.enrichResult(
  x,
  showCategory = 30,
  color = "p.adjust",
  nWords = 4,
  nCluster = 5,
  cex_category = 1,
  label_format = 30,
  xlim = c(0, 2),
  fontsize = 4,
  offset = 0.9,
  offset_tiplab = 0.1,
  hclust_method = "ward.D",
  ...
)

treeplot.compareClusterResult(
  x,
  showCategory = 5,
  color = "p.adjust",
  nWords = 4,
  nCluster = 5,
  cex_category = 1,
  split = NULL,
  label_format = 30,
  xlim = NULL,
  fontsize = 4,
  offset = NULL,
  pie = "equal",
  legend_n = 3,
  offset_tiplab = 0.5,
  hclust_method = "ward.D",
  ...
)
}
\arguments{
\item{x}{enrichment result.}

\item{showCategory}{number of enriched terms to display}

\item{color}{variable that used to color enriched terms, e.g. pvalue,
p.adjust or qvalue}

\item{label_format}{a numeric value sets wrap length, alternatively a
custom function to format axis labels.}

\item{...}{additional parameters}

\item{nWords}{the number of words in the cluster tags}

\item{nCluster}{the number of clusters}

\item{cex_category}{number indicating the amount by which plotting category nodes should be scaled relative to the default.}

\item{xlim}{Limits for the x axes}

\item{fontsize}{the size of text, default is 4.}

\item{offset}{distance bar and tree, offset of bar and text from the clade, default is 0.9.}

\item{offset_tiplab}{tiplab offset}

\item{hclust_method}{method of hclust. This should be (an unambiguous abbreviation of) one of "ward.D", 
"ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{split}{separate result by 'category' variable}

\item{pie}{proportion of clusters in the pie chart, one of
'equal' (default) or 'Count'}

\item{legend_n}{number of circle in legend}
}
\value{
ggplot object
}
\description{
Functional grouping tree diagram for enrichment result of 
over-representation test or gene set enrichment analysis.
}
\details{
This function visualizes gene sets as a tree.
Gene sets with high similarity tend to cluster together, making it easier
for interpretation.
}
\examples{
\dontrun{
    library(clusterProfiler)
    library(org.Hs.eg.db)
    library(enrichplot)
    library(GOSemSim)
    library(DOSE)
    data(geneList)
    gene <- names(geneList)[abs(geneList) > 2]
    ego <- enrichGO(gene  = gene,
        universe      = names(geneList),
        OrgDb         = org.Hs.eg.db,
        ont           = "CC",
        pAdjustMethod = "BH",
        pvalueCutoff  = 0.01,
        qvalueCutoff  = 0.05,
        readable      = TRUE)
    d <- godata('org.Hs.eg.db', ont="BP")
    ego2 <- pairwise_termsim(ego, method = "Wang", semData = d)
    treeplot(ego2, showCategory = 30)
   }
}
