% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/upsetplot.R
\docType{methods}
\name{upsetplot}
\alias{upsetplot}
\alias{upsetplot,enrichResult-method}
\alias{upsetplot,enrichResult,ANY-method}
\alias{upsetplot,gseaResult-method}
\alias{upsetplot,gseaResult}
\alias{upsetplot,mnseaResult-method}
\alias{upsetplot,mnseaResult}
\title{upsetplot method}
\usage{
upsetplot(x, ...)

\S4method{upsetplot}{enrichResult}(x, n = 10, ...)

\S4method{upsetplot}{gseaResult}(x, n = 10, ...)

\S4method{upsetplot}{mnseaResult}(x, n = 10, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters}

\item{n}{number of categories to be plotted}

\item{layer}{Optional \code{mnsea} layer. When \code{NULL}, use collapsed scores.}

\item{value}{score summary to display for overlapping features.}

\item{core_enrichment}{logical. Should only core mnsea features be used?}
}
\value{
plot
}
\description{
upsetplot method generics

Upsetplot
}
\examples{
library(DOSE)
data(geneList)
de <- names(geneList)[1:100]
x <- enrichDO(de)
upsetplot(x, 8)
}
\author{
Guangchuang Yu
}
