% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/upsetplot.R
\docType{methods}
\name{upsetplot}
\alias{upsetplot}
\alias{upsetplot,enrichResult-method}
\alias{upsetplot,enrichResult,ANY-method}
\alias{upsetplot,gseaResult-method}
\alias{upsetplot,gseaResult}
\title{upsetplot method}
\usage{
upsetplot(x, ...)

\S4method{upsetplot}{enrichResult}(x, n = 10, ...)

\S4method{upsetplot}{gseaResult}(x, n = 10, ...)
}
\arguments{
\item{x}{object}

\item{...}{additional parameters}

\item{n}{number of categories to be plotted}
}
\value{
plot
}
\description{
upsetplot method generics
}
\examples{
require(DOSE)
data(geneList)
de=names(geneList)[1:100]
x <- enrichDO(de)
upsetplot(x, 8)
}
\author{
Guangchuang Yu
}
